% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation.R
\name{add_outliers}
\alias{add_outliers}
\title{Annotated Violin Plot}
\usage{
add_outliers(variable, categories, cutoff = 3,
fill = par()$bg, col = par()$fg, bars = par()$fg, lwd = par()$lwd,
verbose = FALSE)
}
\arguments{
\item{variable}{continuous variable to to plot on y-axis (numeric or integer).}

\item{categories}{discrete variable to break down groups (factor or string).}

\item{cutoff}{minimum number (default 3L) of standard deviations to report.}

\item{fill}{colour of spots. Scalar applied to all columns or a vector for each category.}

\item{col}{colour of rings or borders. Scalar applied to all columns or a vector for each category.}

\item{bars}{colour of horizontal bars. Scalar applied to all columns or a vector for each category.}

\item{lwd}{thickness of border.}

\item{verbose}{to print logs (defaults to FALSE).}
}
\description{
Annotation to highlight outliers.
}
\details{
Annotate violin plots with outliers
}
\keyword{annotation}
\keyword{graphics}
\keyword{plot}
\keyword{violin}
