% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vioplotx.R
\name{vioplotx}
\alias{vioplotx}
\title{Enhanced Violin Plot}
\usage{
vioplotx(x, ..., range = 1.5, h = NULL, ylim = NULL, names = NULL,
  horizontal = FALSE, col = "grey50", border = "black", lty = 1,
  lwd = 1, rectCol = "black", lineCol = "black", colMed = "white",
  pchMed = 19, at, add = FALSE, wex = 1, drawRect = TRUE, main = NA,
  sub = NA, xlab = NA, ylab = NA)
}
\arguments{
\item{x}{data vector}

\item{...}{additional data vectors}

\item{range}{a factor to calculate the upper/lower adjacent values}

\item{h}{the height for the density estimator, if omit as explained in sm.density, h will be set to an optimum}

\item{ylim}{y limits}

\item{names}{one label, or a vector of labels for the datas must match the number of datas given}

\item{horizontal}{logical. horizontal or vertical violins}

\item{col}{Graphical parameter for fill colour of the violin(s) polygon. NA for no fill colour. If col is a vector, it specifies the colour per violin, and colours are reused if necessary.}

\item{border}{Graphical parameters for the colour of the violin border passed to lines. NA for no border. If border is a vector, it specifies the colour per bean, and colours are reused if necessary.}

\item{lty, lwd}{Graphical parameters for the violin passed to lines and polygon}

\item{rectCol}{Graphical parameters to control fill colour of the box. NA for no fill colour. If col is a vector, it specifies the colour per violin, and colours are reused if necessary.}

\item{lineCol}{Graphical parameters to control fill colour of the box. NA for no border. If border is a vector, it specifies the colour per violin, and colours are reused if necessary.}

\item{colMed}{Graphical parameters to control colour of the median point}

\item{pchMed}{Graphical parameters to control shape of the median point}

\item{at}{position of each violin. Default to 1:n}

\item{add}{logical. if FALSE (default) a new plot is created}

\item{wex}{relative expansion of the violin.}

\item{drawRect}{logical. the box is drawn if TRUE.}

\item{main, sub, xlab, ylab}{graphical parameters passed to plot.}
}
\description{
Produce violin plot(s) of the given (grouped) values with enhanced annotation and colour per group.
}
\examples{
#vioplotx()
}
\keyword{graphics}
\keyword{plot}
\keyword{violin}

