% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vioplotx.R
\name{vioplotx}
\alias{vioplotx}
\title{Enhanced Violin Plot}
\usage{
vioplotx(x, ..., data = NULL, range = 1.5, h = NULL, ylim = NULL,
  names = NULL, horizontal = FALSE, col = "grey50", border = "black",
  lty = 1, lwd = 1, rectCol = "black", lineCol = "black",
  colMed = "white", pchMed = 19, at, add = FALSE, wex = 1,
  drawRect = TRUE, areaEqual = FALSE, main = NA, sub = NA, xlab = NA,
  ylab = NA, na.action = NULL, na.rm = T)
}
\arguments{
\item{x}{data vector}

\item{...}{additional data vectors}

\item{data}{a data.frame (or list) from which the variables in formula should be taken.}

\item{range}{a factor to calculate the upper/lower adjacent values}

\item{h}{the height for the density estimator, if omit as explained in sm.density, h will be set to an optimum}

\item{ylim}{y limits}

\item{names}{one label, or a vector of labels for the datas must match the number of datas given}

\item{horizontal}{logical. horizontal or vertical violins}

\item{col}{Graphical parameter for fill colour of the violin(s) polygon. NA for no fill colour. If col is a vector, it specifies the colour per violin, and colours are reused if necessary.}

\item{border}{Graphical parameters for the colour of the violin border passed to lines. NA for no border. If border is a vector, it specifies the colour per violin, and colours are reused if necessary.}

\item{lty, lwd}{Graphical parameters for the violin passed to lines and polygon}

\item{rectCol}{Graphical parameters to control fill colour of the box. NA for no fill colour. If col is a vector, it specifies the colour per violin, and colours are reused if necessary.}

\item{lineCol}{Graphical parameters to control fill colour of the box. NA for no border. If border is a vector, it specifies the colour per violin, and colours are reused if necessary.}

\item{colMed}{Graphical parameters to control colour of the median point. If colMed is a vector, it specifies the colour per violin.}

\item{pchMed}{Graphical parameters to control shape of the median point. If pchMed is a vector, it specifies the shape per violin.}

\item{at}{position of each violin. Default to 1:n}

\item{add}{logical. if FALSE (default) a new plot is created}

\item{wex}{relative expansion of the violin.  If wex is a vector, it specifies the area/width size per violin and sizes are reused if necessary.}

\item{drawRect}{logical. The box is drawn if TRUE.}

\item{areaEqual}{logical. Density plots checked for equal area if TRUE. wex must be scalar, relative widths of violins depend on area.}

\item{main, sub, xlab, ylab}{graphical parameters passed to plot.}

\item{na.action}{a function which indicates what should happen when the data contain NAs. The default is to ignore missing values in either the response or the group.}

\item{formula}{a formula, such as y ~ grp, where y is a numeric vector of data values to be split into groups according to the grouping variable grp (usually a factor).}

\item{a}{logical value indicating whether NA values should be stripped before the computation proceeds. Defaults to TRUE.}
}
\description{
Produce violin plot(s) of the given (grouped) values with enhanced annotation and colour per group. Builds upon \code{\link[vioplot]{vioplot}} and intended to function in largely the same way, with added customisation possible with colours for each aspect of the violin, boxplot, and separate violins. This supports input of data as a list or formula, being backwards compatible with \code{\link[vioplot]{vioplot}} and taking input in a formula as used for \code{\link[boxplot]{graphics}}.
}
\examples{

#generate example data
data_one <- rnorm(100)
data_two <- rnorm(50, 1, 2)

#generate violin plot with similar functionality to vioplot
vioplotx(data_one, data_two, col="magenta")

#note vioplox defaults to a greyscale plot
vioplotx(data_one, data_two)

#colours can be customised separately, with axis labels, legends, and titles
vioplotx(data_one, data_two, col=c("red","blue"), names=c("data one", "data two"), main="data violin", xlab="data class", ylab="data read")
legend("topleft", fill=c("red","blue"), legend=c("data one", "data two"))

#colours can be customised for the violin fill and border separately
vioplotx(data_one, data_two, col="grey85", border="purple", names=c("data one", "data two"), main="data violin", xlab="data class", ylab="data read")

#colours can also be customised for the boxplot rectange and lines (border and whiskers)
vioplotx(data_one, data_two, col="grey85", rectCol="lightblue", lineCol="blue", border="purple", names=c("data one", "data two"), main="data violin", xlab="data class", ylab="data read")

#these colours can also be customised separately for each violin
vioplotx(data_one, data_two, col=c("skyblue", "plum"), rectCol=c("lightblue", "palevioletred"), lineCol="blue", border=c("royalblue", "purple"), names=c("data one", "data two"), main="data violin", xlab="data class", ylab="data read")

#this applies to any number of violins, given that colours are provided for each
vioplotx(data_one, data_two, rnorm(200, 3, 0.5), rpois(200, 2.5),  rbinom(100, 10, 0.4), col=c("red", "orange", "green", "blue", "violet"), rectCol=c("palevioletred", "peachpuff", "lightgreen", "lightblue", "plum"), lineCol=c("red4", "orangered", "forestgreen", "royalblue", "mediumorchid"), border=c("red4", "orangered", "forestgreen", "royalblue", "mediumorchid"), names=c("data one", "data two", "data three", "data four", "data five"), main="data violin", xlab="data class", ylab="data read")

#The areaEqual parameter scales with width of violins so that they have equal density area (including missing tails) rather than equal maximum width
vioplotx(data_one, data_two, areaEqual=T)
vioplotx(data_one, data_two, areaEqual=T, col=c("skyblue", "plum"), rectCol=c("lightblue", "palevioletred"), lineCol="blue", border=c("royalblue", "purple"), names=c("data one", "data two"), main="data violin", xlab="data class", ylab="data read")
vioplotx(data_one, data_two, rnorm(200, 3, 0.5), rpois(200, 2.5),  rbinom(100, 10, 0.4), areaEqual=T, col=c("red", "orange", "green", "blue", "violet"), rectCol=c("palevioletred", "peachpuff", "lightgreen", "lightblue", "plum"), lineCol=c("red4", "orangered", "forestgreen", "royalblue", "mediumorchid"), border=c("red4", "orangered", "forestgreen", "royalblue", "mediumorchid"), names=c("data one", "data two", "data three", "data four", "data five"), main="data violin", xlab="data class", ylab="data read")

}
\keyword{graphics}
\keyword{plot}
\keyword{violin}

